fix(build): renumber past main's migrations, and split two files the rebase pushed over the cap
main landed its own 0171, 0172 and 0173 while this branch held the same
three numbers. TestEmbeddedMigrationNamespacesLoad caught it as a
duplicate version rather than letting two branches both claim a slot;
mine move to 0175-0177, headers and STATUS references with them.

jobs.go and organization.go crossed the 500-line cap on main's growth
plus this branch's. Both split along a seam rather than at a line count:
the signal producers now register themselves through addSignalJobs, the
shape addGraphJobs already keeps, and archiving an organization moves
out on its own — it is not one row, since the domains it claims and the
types it wears have to retire with it or a dead account keeps answering
the lists they feed.

'signal' became objectSignal beside the existing objectActivity.
